pyproject.toml
[tool.poetry]
name = "promtior_tech_test"
version = "0.1.0"
description = ""
authors = ["Sebastian Guevara, jseb.guevara@gmail.com"]
readme = "README.md"
packages = [
    { include = "app" },
]

[tool.black]
line-length = 120
target-version = ['py311']
skip-string-normalization = false

[tool.poetry.dependencies]
python = "^3.11"
uvicorn = "^0.23.2"
langserve = {extras = ["all"], version = "^0.3.1"}
pydantic = ">=2.7,<3.0"
langchain-openai = "^0.3.18"
python-decouple = "^3.8"
langchain-community = "^0.3.24"
beautifulsoup4 = "^4.13.4"
faiss-cpu = "^1.11.0"
pypdf = "^5.5.0"
langchain-chroma = "^0.2.4"


[tool.poetry.group.dev.dependencies]
langchain-cli = ">=0.0.15"
black = "^25.1.0"
flake8 = "^7.2.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
